// Gerdau/web/src/redux/user/types.ts
export enum UserActionTypes {
  LOGIN_USER = "LOGIN_USER",
  LOGIN_USER_SUCCESS = "LOGIN_USER_SUCCESS",
  LOGIN_USER_ERROR = "LOGIN_USER_ERROR",

  CHANGE_PASSWORD = "CHANGE_PASSWORD",
  CHANGE_PASSWORD_SUCCESS = "CHANGE_PASSWORD_SUCCESS",
  CHANGE_PASSWORD_ERROR = "CHANGE_PASSWORD_ERROR",

  FETCH_USERS = "FETCH_USERS",
  FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS",
  FETCH_USERS_ERROR = "FETCH_USERS_ERROR",

  RESET_STATE = "RESET_STATE",

  FETCH_USER = "FETCH_USER",
  FETCH_USER_SUCCESS = "FETCH_USER_SUCCESS",
  FETCH_USER_ERROR = "FETCH_USER_ERROR",

  // REGISTER_USER = 'REGISTER_USER',
  // REGISTER_USER_SUCCESS = 'REGISTER_USER_SUCCESS',
  // REGISTER_USER_ERROR = 'REGISTER_USER_ERROR',

  // RESET_USER = 'RESET_USER',
  // RESET_USER_SUCCESS = 'RESET_USER_SUCCESS',
  // RESET_USER_ERROR = 'RESET_USER_ERROR',

  // VALIDATE_TOKEN = 'VALIDATE_TOKEN',
  // VALIDATE_TOKEN_SUCCESS = 'VALIDATE_TOKEN_SUCCESS',
  // VALIDATE_TOKEN_ERROR = 'VALIDATE_TOKEN_ERROR',

  // RESEND_CONFIRMATION = 'RESEND_CONFIRMATION',
  // RESEND_CONFIRMATION_SUCCESS = 'RESEND_CONFIRMATION_SUCCESS',
  // RESEND_CONFIRMATION_ERROR = 'RESEND_CONFIRMATION_ERROR',

  // RESET_STATE = 'RESET_STATE',
}